Add C++ implementation for ApproShapley
Implement a C++ version for the ApproShapley algorithm. The glove
game is used as an example game.
Measure execution times with and without threads.

This implementation is just an experiment. Some edge cases are not
yet considered (e.g. integer overflow).

We get some good speed up by using C++ for this specific algorithm
although one should note, that depending on the characteristic
function, the c++ compiler can do a better or worse job with its
optimizations. If the characteristic function is later defined in
R and we have to call it inside our C++ code I do not know how
much of a speed up we are going to get...