Add sample_from_partition in C++
Add sample_from_partition function for C++. This is useful to get
a sample that is compatible with a certain partition.
If this sample function is used instead of std::shuffle to get an
order, the existing code should be able to work with a priori
unions.
(Out of "Estimation of the Owen Value Based on Sampling" - 2018)